Removed tutorial related classes and calls; simplified deploy logic.
This still implements #2. Now the logic for the deploy operation is
simplified: the user can’t specify a name in the deploy URI anymore. It
is assumed that the name of the driver contained in the archive file’s
`drivers` folder is the same as the name of the archive itself.